requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
click==7.1.2
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
flake8==3.8.3
flake8-builtins==1.5.3
flake8-docstrings==1.5.0
flake8-import-order==0.18.1
flake8-quotes==3.2.0
importlib-metadata==1.7.0
Markdown==3.2.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pytz==2020.1
regex==2020.7.14
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
zipp==3.1.0
